requirements.txt
amqp==5.1.1
anyio==4.0.0
asgiref==3.7.2
asttokens==2.2.1
async-timeout==4.0.3
backcall==0.2.0
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
decorator==4.4.2
Django==4.2.6
django-debug-toolbar==4.2.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
emoji==2.8.0
executing==1.2.0
gunicorn==21.2.0
h11==0.14.0
httptools==0.6.0
idna==3.4
imageio==2.31.5
imageio-ffmpeg==0.4.9
ipython==8.11.0
jedi==0.18.2
kombu==5.3.2
matplotlib-inline==0.1.6
moviepy==1.0.3
nest-asyncio==1.5.8
numpy==1.26.0
packaging==23.2
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.1
pilmoji==2.0.4
proglog==0.1.10
prompt-toolkit==3.0.39
pure-eval==0.2.2
Pygments==2.14.0
pyngrok==7.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==4.6.0
requests==2.31.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
stack-data==0.6.2
tqdm==4.66.1
traitlets==5.9.0
tzdata==2023.3
urllib3==2.0.6
uvicorn==0.23.2
vine==5.0.0
watchfiles==0.20.0
wcwidth==0.2.8
websockets==11.0.3
whitenoise==6.0.0
